Refactor: move functions to single file for org / cleaner imports
Database seed files became bloated with imports so I moved all these
similar functions to a single file to clean up imports.
